requirements.txt
# AI Fusion - 智能多模型融合系统
# 基于PocketFlow的异步LLM编排框架

# ============================================
# 核心框架
# ============================================
pocketflow>=0.0.3

# ============================================
# LLM API客户端
# ============================================
openai>=1.3.0
anthropic>=0.7.0
httpx>=0.28.0

# ============================================
# 异步HTTP和网络
# ============================================
aiohttp>=3.8.0
requests>=2.28.0
anyio>=4.0.0
fastapi>=0.110.0
uvicorn>=0.24.0

# ============================================
# 环境配置
# ============================================
python-dotenv>=1.0.0

# ============================================
# 数据处理和类型支持
# ============================================
typing-extensions>=4.0.0
pydantic>=2.0.0

# ============================================
# 开发和测试工具（可选）
# ============================================
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
# ============================================
# Python版本要求
# ============================================
# Python >= 3.8
